Make a geo table its own DC type flavour instead of a capability
A separate Capabilities column only ever held one badge, and "table + Geo"
described the same thing twice. Coordinates belong on the type axis: a table
with lat/lon is a Geomap, which is exactly what Map components bind to.

- dcTypeIcon owns the type axis end to end: the icon/label/colour map now
  carries a geomap entry, plus the coordinates predicate and the lat/lon
  column reader that used to sit in ProjectDetailApp. dcTypeMeta takes a geo
  flag, dcTypeMetaFor resolves it from a whole collection, and DcTypeIcon
  takes the same flag, so a geomap shows the grape marker glyph wherever it
  appears. Callers holding only a type string (the ingestion report reads a
  report payload, not a config) are unaffected.
- dcBadges becomes dcMetatype and keeps only the metatype axis.
- Manager table: the Capabilities column is gone and the Type column names the
  flavour (Table, Geomap, MultiQC, Image, GeoJSON, Phylo tree) instead of
  echoing the raw dc_type. Filtering and sorting follow the label shown, so
  typing "geomap" finds the geo tables.
- DC viewer: the title glyph is the shared type icon, the Type row reads
  GEOMAP in grape with a tooltip spelling out that the stored type is a table
  carrying lat/lon, and the Capabilities row is gone. The lat/lon rows stay
  where they were.
- Phylogeny is labelled "Phylo tree".

Not verified in a running instance: no local stack was up.
